test(e2e): suites Playwright désambiguïsées + fusion login + script staging npm (Closes #6592)
- e2e/navigation.spec.js → navigation-spa.spec.js ; e2e-staging/navigation.spec.js
  → navigation-blade.spec.js (2 suites homonymes 100 % divergentes).
- login-smoke.spec.js fusionne login-smoke + navigation-smoke (assertions
  dédupliquées : structure formulaire, validation vide, identifiants invalides,
  deep-link #2334) ; navigation-smoke.spec.js supprimé (redirect non-auth
  déjà couvert par navigation-spa).
- package.json : script npm run test:e2e:staging ; CI e2e-staging.yml câblée
  dessus (parité locale-CI).
- Vérifié : 9/9 specs SPA verts (login-smoke 4 + navigation-spa 5), eslint OK.
