+++ .github/workflows/e2e-staging.yml
@@ -254,7 +254,7 @@ jobs:
         env:
           CI: true
           BASE_URL: ${{ github.event.inputs.staging_url || env.DEFAULT_STAGING_URL }}
-        run: npx playwright test --config=playwright.staging.config.js
+        run: npm run test:e2e:staging
 
       - name: Upload Playwright report
         if: always()
