+++ front/admin-dashboard/e2e/navigation-spa.spec.js
