+++ front/admin-dashboard/package.json
@@ -10,7 +10,8 @@
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "format": "prettier --write src/",
-    "test:e2e": "playwright test"
+    "test:e2e": "playwright test",
+    "test:e2e:staging": "playwright test --config playwright.staging.config.js"
   },
   "dependencies": {
     "@headlessui/vue": "^1.7.16",
